Added API for user-defined code to Prioritizer, changed crawl priority from Int to Long

+++ scripts/abwcf-schema-postgres.sql
@@ -2,7 +2,7 @@ CREATE TABLE pages (
   url            TEXT,
   status         TEXT,
   crawl_depth    INTEGER,
-  crawl_priority INTEGER
+  crawl_priority BIGINT
 );
 
 -- Indexes to speed up common ABWCF queries:
